Add unsaved logic to custom hook
