+++ src/components/blog/EditPostForm.tsx
@@ -5,6 +5,7 @@ import { useRouter } from 'next/navigation';
 import { updatePost, deletePost } from '@/services/post-service';
 import { useToast } from '@/components/ui/ToastContext';
 import { useImageInsertion } from '@/hooks/useImageInsertion';
+import { useUnsavedChanges } from '@/hooks/useUnsavedChanges';
 import { useEditPostStore } from '@/store/edit';
 import { getCategories } from '@/services/category-service';
 
@@ -13,7 +14,6 @@ import Button from '@/components/ui/Button';
 import Modal from '@/components/ui/Modal';
 import MediaGallery from '@/components/blog/MediaGallery';
 import PostFormControls from '@/components/blog/PostFormControls';
-import UnsavedChangesHandler from '@/components/ui/UnsavedChangesHandler';
 
 import { BlogPost, Category } from '@/types/blog';
 
@@ -158,6 +158,15 @@ export default function EditPostForm({
   }, []);
 
   const {
+    showUnsavedChangesModal,
+    handleConfirmNavigation,
+    handleCancelNavigation,
+  } = useUnsavedChanges({
+    hasUnsavedChanges,
+    onDiscard: discardChanges,
+  });
+
+  const {
     textareaRef,
     cursorPosition,
     handleTextAreaSelect,
@@ -259,9 +268,13 @@ export default function EditPostForm({
         />
       </Modal>
 
-      <UnsavedChangesHandler
-        hasUnsavedChanges={hasUnsavedChanges}
-        onDiscard={discardChanges}
+      <Modal
+        isOpen={showUnsavedChangesModal}
+        title="Unsaved Changes"
+        message="You have unsaved changes. Are you sure you want to leave this page?"
+        onCancel={handleCancelNavigation}
+        onConfirm={handleConfirmNavigation}
+        confirmText="Leave Page"
       />
     </>
   );
